Move test to model.py
